HDPI-6407: Align view building components to match an interface and naming convention
