+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/view/AlternativesToPossessionView.java
@@ -14,8 +14,9 @@ import java.util.Optional;
 import static uk.gov.hmcts.reform.pcs.ccd.util.YesOrNoConverter.toVerticalYesNo;
 
 @Component
-public class AlternativesToPossessionView {
+public class AlternativesToPossessionView implements ViewComponent {
 
+    @Override
     public void setCaseFields(PCSCase pcsCase, PcsCaseEntity pcsCaseEntity) {
         getMainClaim(pcsCaseEntity)
             .map(ClaimEntity::getPossessionAlternativesEntity)
