+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/view/ClaimGroundsView.java
@@ -31,10 +31,11 @@ import java.util.Optional;
 import java.util.Set;
 
 @Component
-public class ClaimGroundsView {
+public class ClaimGroundsView implements ViewComponent {
 
     private static final Comparator<ListValue<ClaimGroundSummary>> GROUNDS_COMPARATOR = createGroundsComparator();
 
+    @Override
     public void setCaseFields(PCSCase pcsCase, PcsCaseEntity pcsCaseEntity) {
         getMainClaim(pcsCaseEntity)
             .ifPresent(mainClaim -> setClaimGroundFields(pcsCase, mainClaim));
