+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/view/NoticeOfPossessionView.java
@@ -13,8 +13,9 @@ import uk.gov.hmcts.reform.pcs.postcodecourt.model.LegislativeCountry;
 import java.util.Optional;
 
 @Component
-public class NoticeOfPossessionView {
+public class NoticeOfPossessionView implements ViewComponent {
 
+    @Override
     public void setCaseFields(PCSCase pcsCase, PcsCaseEntity pcsCaseEntity) {
         getMainClaim(pcsCaseEntity)
             .map(ClaimEntity::getNoticeOfPossession)
