+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/view/RentArrearsView.java
@@ -15,8 +15,9 @@ import java.util.Optional;
 import java.util.Set;
 
 @Component
-public class RentArrearsView {
+public class RentArrearsView implements ViewComponent {
 
+    @Override
     public void setCaseFields(PCSCase pcsCase, PcsCaseEntity pcsCaseEntity) {
         getMainClaim(pcsCaseEntity)
             .map(ClaimEntity::getRentArrears)
