+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/view/RentDetailsView.java
@@ -7,8 +7,9 @@ import uk.gov.hmcts.reform.pcs.ccd.entity.PcsCaseEntity;
 import uk.gov.hmcts.reform.pcs.ccd.entity.TenancyLicenceEntity;
 
 @Component
-public class RentDetailsView {
+public class RentDetailsView implements ViewComponent {
 
+    @Override
     public void setCaseFields(PCSCase pcsCase, PcsCaseEntity pcsCaseEntity) {
         TenancyLicenceEntity tenancyLicence = pcsCaseEntity.getTenancyLicence();
