+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/view/TenancyLicenceView.java
@@ -12,8 +12,9 @@ import uk.gov.hmcts.reform.pcs.ccd.entity.TenancyLicenceEntity;
 import uk.gov.hmcts.reform.pcs.postcodecourt.model.LegislativeCountry;
 
 @Component
-public class TenancyLicenceView {
+public class TenancyLicenceView implements ViewComponent {
 
+    @Override
     public void setCaseFields(PCSCase pcsCase, PcsCaseEntity pcsCaseEntity) {
         TenancyLicenceEntity tenancyLicence = pcsCaseEntity.getTenancyLicence();
