+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/view/ViewComponent.java
@@ -1,0 +1,10 @@
+package uk.gov.hmcts.reform.pcs.ccd.view;
+
+import uk.gov.hmcts.reform.pcs.ccd.domain.PCSCase;
+import uk.gov.hmcts.reform.pcs.ccd.entity.PcsCaseEntity;
+
+public interface ViewComponent {
+
+    void setCaseFields(PCSCase pcsCase, PcsCaseEntity pcsCaseEntity);
+
+}
