+++ src/test/java/uk/gov/hmcts/reform/pcs/ccd/PCSCaseViewTest.java
@@ -9,42 +9,35 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import org.modelmapper.ModelMapper;
 import uk.gov.hmcts.ccd.sdk.CaseViewRequest;
 import uk.gov.hmcts.ccd.sdk.type.AddressUK;
-import uk.gov.hmcts.ccd.sdk.type.ListValue;
 import uk.gov.hmcts.reform.pcs.ccd.domain.PCSCase;
-import uk.gov.hmcts.reform.pcs.ccd.domain.Party;
 import uk.gov.hmcts.reform.pcs.ccd.domain.State;
-import uk.gov.hmcts.reform.pcs.ccd.enforcementorder.EnforcementOrderMediator;
 import uk.gov.hmcts.reform.pcs.ccd.entity.AddressEntity;
 import uk.gov.hmcts.reform.pcs.ccd.entity.ClaimEntity;
-import uk.gov.hmcts.reform.pcs.ccd.entity.DocumentEntity;
 import uk.gov.hmcts.reform.pcs.ccd.entity.PcsCaseEntity;
-import uk.gov.hmcts.reform.pcs.ccd.entity.party.ClaimPartyEntity;
-import uk.gov.hmcts.reform.pcs.ccd.entity.party.ClaimPartyId;
-import uk.gov.hmcts.reform.pcs.ccd.entity.party.PartyEntity;
-import uk.gov.hmcts.reform.pcs.ccd.entity.party.PartyRole;
 import uk.gov.hmcts.reform.pcs.ccd.repository.PcsCaseRepository;
 import uk.gov.hmcts.reform.pcs.ccd.service.CaseTitleService;
 import uk.gov.hmcts.reform.pcs.ccd.service.DraftCaseDataService;
 import uk.gov.hmcts.reform.pcs.ccd.view.AlternativesToPossessionView;
 import uk.gov.hmcts.reform.pcs.ccd.view.AsbProhibitedConductView;
+import uk.gov.hmcts.reform.pcs.ccd.view.CaseLinkView;
 import uk.gov.hmcts.reform.pcs.ccd.view.ClaimGroundsView;
 import uk.gov.hmcts.reform.pcs.ccd.view.ClaimView;
+import uk.gov.hmcts.reform.pcs.ccd.view.DocumentsView;
 import uk.gov.hmcts.reform.pcs.ccd.view.HousingActWalesView;
 import uk.gov.hmcts.reform.pcs.ccd.view.NoticeOfPossessionView;
+import uk.gov.hmcts.reform.pcs.ccd.view.PartyView;
 import uk.gov.hmcts.reform.pcs.ccd.view.RentArrearsView;
 import uk.gov.hmcts.reform.pcs.ccd.view.RentDetailsView;
 import uk.gov.hmcts.reform.pcs.ccd.view.StatementOfTruthView;
 import uk.gov.hmcts.reform.pcs.ccd.view.TenancyLicenceView;
+import uk.gov.hmcts.reform.pcs.ccd.view.enforcementorder.EnforcementOrderView;
 import uk.gov.hmcts.reform.pcs.ccd.view.globalsearch.CaseFieldsView;
-import uk.gov.hmcts.reform.pcs.ccd.view.CaseLinkView;
 import uk.gov.hmcts.reform.pcs.exception.CaseNotFoundException;
 import uk.gov.hmcts.reform.pcs.postcodecourt.model.LegislativeCountry;
 import uk.gov.hmcts.reform.pcs.security.SecurityContextService;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -74,6 +67,10 @@ class PCSCaseViewTest {
     @Mock
     private ClaimView claimView;
     @Mock
+    private PartyView partyView;
+    @Mock
+    private DocumentsView documentsView;
+    @Mock
     private TenancyLicenceView tenancyLicenceView;
     @Mock
     private ClaimGroundsView claimGroundsView;
@@ -99,7 +96,7 @@ class PCSCaseViewTest {
     @Mock
     private CaseFieldsView caseFieldsView;
     @Mock
-    private EnforcementOrderMediator enforcementOrderMediator;
+    private EnforcementOrderView enforcementOrderView;
     @Mock
     private CaseLinkView caseLinkView;
 
@@ -111,11 +108,11 @@ class PCSCaseViewTest {
         when(pcsCaseEntity.getClaims()).thenReturn(List.of(claimEntity));
 
         underTest = new PCSCaseView(pcsCaseRepository, securityContextService, modelMapper, draftCaseDataService,
-                                    caseTitleService, claimView, tenancyLicenceView, claimGroundsView, rentDetailsView,
-                                    alternativesToPossessionView, housingActWalesView, asbProhibitedConductView,
-                                    rentArrearsView, noticeOfPossessionView,
-                                    statementOfTruthView, caseFieldsView, caseLinkView, enforcementOrderMediator
-        );
+                                    caseTitleService, claimView, partyView, documentsView, tenancyLicenceView,
+                                    claimGroundsView, rentDetailsView, alternativesToPossessionView,
+                                    housingActWalesView, asbProhibitedConductView, rentArrearsView,
+                                    noticeOfPossessionView, statementOfTruthView, caseFieldsView, caseLinkView,
+                                    enforcementOrderView);
     }
 
     @Test
@@ -177,25 +174,6 @@ class PCSCaseViewTest {
     }
 
     @Test
-    void shouldMapPartyEntity() {
-        // Given
-        PartyEntity partyEntity = mock(PartyEntity.class);
-        when(pcsCaseEntity.getParties()).thenReturn(Set.of(partyEntity));
-
-        Party party = mock(Party.class);
-
-        when(modelMapper.map(partyEntity, Party.class)).thenReturn(party);
-
-        // When
-        PCSCase pcsCase = underTest.getCase(request(CASE_REFERENCE, DEFAULT_STATE));
-
-        // Then
-        List<ListValue<Party>> mappedParties = pcsCase.getParties();
-        assertThat(mappedParties).hasSize(1);
-        assertThat(mappedParties.getFirst().getValue()).isSameAs(party);
-    }
-
-    @Test
     void shouldMapLegislativeCountry() {
         // Given
         LegislativeCountry expectedLegislativeCountry = LegislativeCountry.SCOTLAND;
@@ -209,126 +187,14 @@ class PCSCaseViewTest {
     }
 
     @Test
-    void shouldMapAllParties() {
-        // Given
-        Party claimant = mock(Party.class);
-        UUID claimantId = UUID.randomUUID();
-        ClaimPartyEntity claimantClaimParty = createClaimPartyEntity(claimant, claimantId, PartyRole.CLAIMANT);
-
-        Party defendant1 = mock(Party.class);
-        UUID defendant1Id = UUID.randomUUID();
-        ClaimPartyEntity defendant1ClaimParty = createClaimPartyEntity(defendant1, defendant1Id, PartyRole.DEFENDANT);
-
-        Party defendant2 = mock(Party.class);
-        UUID defendant2Id = UUID.randomUUID();
-        ClaimPartyEntity defendant2ClaimParty = createClaimPartyEntity(defendant2, defendant2Id, PartyRole.DEFENDANT);
-
-        Party underlessee1 = mock(Party.class);
-        UUID underlessee1Id = UUID.randomUUID();
-        ClaimPartyEntity underlessee1ClaimParty = createClaimPartyEntity(
-            underlessee1,
-            underlessee1Id,
-            PartyRole.UNDERLESSEE_OR_MORTGAGEE
-        );
-
-        Party underlessee2 = mock(Party.class);
-        UUID underlessee2Id = UUID.randomUUID();
-        ClaimPartyEntity underlessee2ClaimParty = createClaimPartyEntity(
-            underlessee2,
-            underlessee2Id,
-            PartyRole.UNDERLESSEE_OR_MORTGAGEE
-        );
-
-        when(claimEntity.getClaimParties()).thenReturn(
-            List.of(claimantClaimParty, defendant1ClaimParty, defendant2ClaimParty,
-                    underlessee1ClaimParty, underlessee2ClaimParty
-            ));
-
-        // When
-        PCSCase pcsCase = underTest.getCase(request(CASE_REFERENCE, DEFAULT_STATE));
-
-        // Then
-        assertThat(pcsCase.getAllClaimants())
-            .containsExactly(asListValue(claimantId, claimant));
-
-        assertThat(pcsCase.getAllDefendants())
-            .containsExactly(
-                asListValue(defendant1Id, defendant1),
-                asListValue(defendant2Id, defendant2)
-            );
-
-        assertThat(pcsCase.getAllUnderlesseeOrMortgagees())
-            .containsExactly(
-                asListValue(underlessee1Id, underlessee1),
-                asListValue(underlessee2Id, underlessee2)
-            );
-
-    }
-
-    @Test
-    void shouldReturnEmptyListWhenNoDocumentsExist() {
-        // Given
-        when(pcsCaseEntity.getDocuments()).thenReturn(List.of());
-
-        // When
-        PCSCase pcsCase = underTest.getCase(request(CASE_REFERENCE, DEFAULT_STATE));
-
-        // Then
-        assertThat(pcsCase.getAllDocuments()).isEmpty();
-    }
-
-    @Test
-    void shouldMapDocuments() {
-        // Given
-        DocumentEntity entity1 = DocumentEntity.builder()
-            .id(UUID.randomUUID())
-            .fileName("doc1.pdf")
-            .url("url1")
-            .build();
-
-        DocumentEntity entity2 = DocumentEntity.builder()
-            .id(UUID.randomUUID())
-            .fileName("doc2.pdf")
-            .url("url2")
-            .build();
-
-        when(pcsCaseEntity.getDocuments()).thenReturn(List.of(entity1,entity2));
-
-        // When
-        PCSCase pcsCase = underTest.getCase(request(CASE_REFERENCE, DEFAULT_STATE));
-
-        //Then
-        assertThat(pcsCase.getAllDocuments()).hasSize(2);
-        assertThat(pcsCase.getAllDocuments()).extracting(lv -> lv.getValue().getFilename())
-            .containsExactly("doc1.pdf", "doc2.pdf");
-    }
-
-    private static ListValue<Party> asListValue(UUID id, Party party) {
-        return ListValue.<Party>builder().id(id.toString()).value(party).build();
-    }
-
-    private ClaimPartyEntity createClaimPartyEntity(Party party, UUID partyId, PartyRole partyRole) {
-        PartyEntity partyEntity = mock(PartyEntity.class);
-
-        when(modelMapper.map(partyEntity, Party.class)).thenReturn(party);
-
-        ClaimPartyId claimPartyId = new ClaimPartyId();
-        claimPartyId.setPartyId(partyId);
-
-        return ClaimPartyEntity.builder()
-            .id(claimPartyId)
-            .role(partyRole)
-            .party(partyEntity)
-            .build();
-    }
-
-    @Test
     void shouldSetCaseFieldsInViewHelpers() {
         // When
         PCSCase pcsCase = underTest.getCase(request(CASE_REFERENCE, DEFAULT_STATE));
 
         // Then
         verify(claimView).setCaseFields(pcsCase, pcsCaseEntity);
+        verify(partyView).setCaseFields(pcsCase, pcsCaseEntity);
+        verify(documentsView).setCaseFields(pcsCase, pcsCaseEntity);
         verify(tenancyLicenceView).setCaseFields(pcsCase, pcsCaseEntity);
         verify(claimGroundsView).setCaseFields(pcsCase, pcsCaseEntity);
         verify(rentDetailsView).setCaseFields(pcsCase, pcsCaseEntity);
@@ -339,6 +205,7 @@ class PCSCaseViewTest {
         verify(noticeOfPossessionView).setCaseFields(pcsCase, pcsCaseEntity);
         verify(statementOfTruthView).setCaseFields(pcsCase, pcsCaseEntity);
         verify(caseLinkView).setCaseFields(pcsCase, pcsCaseEntity);
+        verify(enforcementOrderView).setCaseFields(pcsCase, pcsCaseEntity);
     }
 
     @Test
@@ -350,15 +217,6 @@ class PCSCaseViewTest {
 
         // Then
         verify(caseFieldsView).setCaseFields(pcsCase);
-    }
-
-    @Test
-    void shouldCallEnforcementOrderMediator() {
-        // When
-        PCSCase pcsCase = underTest.getCase(request(CASE_REFERENCE, DEFAULT_STATE));
-
-        // Then
-        verify(enforcementOrderMediator).handleEnforcementRequirements(CASE_REFERENCE, pcsCase);
     }
 
     private AddressUK stubAddressEntityModelMapper(AddressEntity addressEntity) {
