+++ src/test/java/uk/gov/hmcts/reform/pcs/ccd/view/enforcementorder/EnforcementOrderViewTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.ccd.enforcementorder;
+package uk.gov.hmcts.reform.pcs.ccd.view.enforcementorder;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -15,9 +15,7 @@ import uk.gov.hmcts.reform.pcs.ccd.domain.enforcetheorder.EnforcementOrder;
 import uk.gov.hmcts.reform.pcs.ccd.entity.ClaimEntity;
 import uk.gov.hmcts.reform.pcs.ccd.entity.PcsCaseEntity;
 import uk.gov.hmcts.reform.pcs.ccd.entity.enforcetheorder.EnforcementOrderEntity;
-import uk.gov.hmcts.reform.pcs.ccd.repository.PcsCaseRepository;
 import uk.gov.hmcts.reform.pcs.ccd.repository.enforcetheorder.EnforcementOrderRepository;
-import uk.gov.hmcts.reform.pcs.exception.CaseNotFoundException;
 
 import java.time.LocalDateTime;
 import java.time.ZoneId;
@@ -33,26 +31,19 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class EnforcementOrderMediatorTest {
+class EnforcementOrderViewTest {
 
-    private static final long CASE_REFERENCE = 1234567890L;
-
-    @Mock
-    private PcsCaseRepository pcsCaseRepository;
     @Mock
     private EnforcementOrderRepository enforcementOrderRepository;
-
     @InjectMocks
-    private EnforcementOrderMediator underTest;
+    private EnforcementOrderView underTest;
 
     private PCSCase pcsCase;
 
@@ -66,65 +57,55 @@ class EnforcementOrderMediatorTest {
     @Test
     void shouldSetShowConfirmEvictionJourneyToYesWhenBailiffDateExists() {
         // Given
-        long caseReference = 1234567890L;
         LocalDateTime bailiffDate = LocalDateTime.parse("2026-04-15T10:00:00");
 
         PcsCaseEntity pcsCaseEntity = createPcsCaseEntity();
         EnforcementOrderEntity enforcementOrderEntity = createEnforcementOrderEntity(bailiffDate);
 
-        when(pcsCaseRepository.findByCaseReference(caseReference))
-            .thenReturn(Optional.of(pcsCaseEntity));
         when(enforcementOrderRepository.findByClaimId(any(UUID.class)))
             .thenReturn(List.of(enforcementOrderEntity));
 
         // When
-        underTest.handleEnforcementRequirements(caseReference, pcsCase);
+        underTest.setCaseFields(pcsCase, pcsCaseEntity);
 
         // Then
         assertEquals(YesOrNo.YES, pcsCase.getShowConfirmEvictionJourney());
         assertNotNull(pcsCase.getConfirmEvictionSummaryMarkup());
         assertTrue(pcsCase.getConfirmEvictionSummaryMarkup().contains("Confirm the eviction date"));
-        verify(pcsCaseRepository).findByCaseReference(caseReference);
         verify(enforcementOrderRepository).findByClaimId(any(UUID.class));
     }
 
     @Test
     void shouldSetShowConfirmEvictionJourneyToNoWhenBailiffDateIsNull() {
         // Given
-        long caseReference = 1234567890L;
         PcsCaseEntity pcsCaseEntity = createPcsCaseEntity();
         EnforcementOrderEntity enforcementOrderEntity = createEnforcementOrderEntity(null);
 
-        when(pcsCaseRepository.findByCaseReference(caseReference))
-            .thenReturn(Optional.of(pcsCaseEntity));
         when(enforcementOrderRepository.findByClaimId(any(UUID.class)))
             .thenReturn(List.of(enforcementOrderEntity));
 
         // When
-        underTest.handleEnforcementRequirements(caseReference, pcsCase);
+        underTest.setCaseFields(pcsCase, pcsCaseEntity);
 
         // Then
         assertEquals(YesOrNo.NO, pcsCase.getShowConfirmEvictionJourney());
         assertNotNull(pcsCase.getConfirmEvictionSummaryMarkup());
         assertTrue(pcsCase.getConfirmEvictionSummaryMarkup().contains("You cannot enforce the order at the moment"));
-        verify(pcsCaseRepository).findByCaseReference(caseReference);
         verify(enforcementOrderRepository).findByClaimId(any(UUID.class));
     }
 
     @Test
     void shouldFormatBailiffDateCorrectlyInMarkup() {
         // Given
-        long caseReference = 1234567890L;
         LocalDateTime bailiffDate = LocalDateTime.parse("2026-05-20T14:30:00");
 
         PcsCaseEntity pcsCaseEntity = createPcsCaseEntity();
         EnforcementOrderEntity enforcementOrderEntity = createEnforcementOrderEntity(bailiffDate);
 
-        when(pcsCaseRepository.findByCaseReference(caseReference)).thenReturn(Optional.of(pcsCaseEntity));
         when(enforcementOrderRepository.findByClaimId(any(UUID.class))).thenReturn(List.of(enforcementOrderEntity));
 
         // When
-        underTest.handleEnforcementRequirements(caseReference, pcsCase);
+        underTest.setCaseFields(pcsCase, pcsCaseEntity);
 
         // Then
         DateTimeFormatter outputFormatter = DateTimeFormatter.ofPattern("EEEE, d MMMM yyyy", Locale.UK);
@@ -135,19 +116,16 @@ class EnforcementOrderMediatorTest {
     @Test
     void shouldCalculateDeadlineDateAsMinus72HoursFromBailiffDate() {
         // Given
-        long caseReference = 1234567890L;
         LocalDateTime bailiffDate = LocalDateTime.parse("2026-05-20T14:30:00");
 
         PcsCaseEntity pcsCaseEntity = createPcsCaseEntity();
         EnforcementOrderEntity enforcementOrderEntity = createEnforcementOrderEntity(bailiffDate);
 
-        when(pcsCaseRepository.findByCaseReference(caseReference))
-            .thenReturn(Optional.of(pcsCaseEntity));
         when(enforcementOrderRepository.findByClaimId(any(UUID.class)))
             .thenReturn(List.of(enforcementOrderEntity));
 
         // When
-        underTest.handleEnforcementRequirements(caseReference, pcsCase);
+        underTest.setCaseFields(pcsCase, pcsCaseEntity);
 
         // Then
         DateTimeFormatter outputFormatter = DateTimeFormatter.ofPattern("d MMMM yyyy", Locale.UK);
@@ -156,117 +134,48 @@ class EnforcementOrderMediatorTest {
     }
 
     @Test
-    void shouldNotProcessWhenCaseReferenceIsZero() {
-        // Given
-        long caseReference = 0L;
-
-        // When
-        underTest.handleEnforcementRequirements(caseReference, pcsCase);
-
-        // Then
-        assertNull(pcsCase.getShowConfirmEvictionJourney());
-        assertNull(pcsCase.getConfirmEvictionSummaryMarkup());
-        verify(pcsCaseRepository, never()).findByCaseReference(anyLong());
-    }
-
-    @Test
-    void shouldNotProcessWhenCaseReferenceIsNegative() {
-        // Given
-        long caseReference = -1L;
-
-        // When
-        underTest.handleEnforcementRequirements(caseReference, pcsCase);
-
-        // Then
-        assertNull(pcsCase.getShowConfirmEvictionJourney());
-        assertNull(pcsCase.getConfirmEvictionSummaryMarkup());
-        verify(pcsCaseRepository, never()).findByCaseReference(anyLong());
-    }
-
-    @Test
-    void shouldNotProcessWhenPcsCaseIsNull() {
-        // Given
-        long caseReference = 1234567890L;
-
-        // When
-        underTest.handleEnforcementRequirements(caseReference, null);
-
-        // Then
-        verify(pcsCaseRepository, never()).findByCaseReference(anyLong());
-    }
-
-    @Test
-    void shouldThrowCaseNotFoundExceptionWhenCaseDoesNotExist() {
-        // Given
-        long caseReference = 1234567890L;
-
-        when(pcsCaseRepository.findByCaseReference(caseReference))
-            .thenReturn(Optional.empty());
-
-        // When & Then
-        assertThrows(CaseNotFoundException.class, () ->
-            underTest.handleEnforcementRequirements(caseReference, pcsCase)
-        );
-        verify(pcsCaseRepository).findByCaseReference(caseReference);
-    }
-
-    @Test
     void shouldReturnEmptyWhenClaimsListIsNull() {
         // Given
-        long caseReference = 1234567890L;
         PcsCaseEntity pcsCaseEntity = new PcsCaseEntity();
         pcsCaseEntity.setClaims(null);
 
-        when(pcsCaseRepository.findByCaseReference(caseReference))
-            .thenReturn(Optional.of(pcsCaseEntity));
-
         // When
-        Optional<EnforcementOrderEntity> result = underTest.getEnforcementOrder(caseReference);
+        Optional<EnforcementOrderEntity> result = underTest.getEnforcementOrder(pcsCaseEntity);
 
         // Then
         assertTrue(result.isEmpty());
-        verify(pcsCaseRepository).findByCaseReference(caseReference);
         verify(enforcementOrderRepository, never()).findByClaimId(any());
     }
 
     @Test
     void shouldReturnEmptyWhenClaimsListIsEmpty() {
         // Given
-        long caseReference = 1234567890L;
         PcsCaseEntity pcsCaseEntity = new PcsCaseEntity();
         pcsCaseEntity.setClaims(new ArrayList<>());
 
-        when(pcsCaseRepository.findByCaseReference(caseReference))
-            .thenReturn(Optional.of(pcsCaseEntity));
-
         // When
-        Optional<EnforcementOrderEntity> result = underTest.getEnforcementOrder(caseReference);
+        Optional<EnforcementOrderEntity> result = underTest.getEnforcementOrder(pcsCaseEntity);
 
         // Then
         assertTrue(result.isEmpty());
-        verify(pcsCaseRepository).findByCaseReference(caseReference);
         verify(enforcementOrderRepository, never()).findByClaimId(any());
     }
 
     @Test
     void shouldReturnEnforcementOrderWhenFound() {
         // Given
-        long caseReference = 1234567890L;
         PcsCaseEntity pcsCaseEntity = createPcsCaseEntity();
         EnforcementOrderEntity enforcementOrderEntity = createEnforcementOrderEntity(LocalDateTime.now());
 
-        when(pcsCaseRepository.findByCaseReference(caseReference))
-            .thenReturn(Optional.of(pcsCaseEntity));
         when(enforcementOrderRepository.findByClaimId(any(UUID.class)))
             .thenReturn(List.of(enforcementOrderEntity));
 
         // When
-        Optional<EnforcementOrderEntity> result = underTest.getEnforcementOrder(caseReference);
+        Optional<EnforcementOrderEntity> result = underTest.getEnforcementOrder(pcsCaseEntity);
 
         // Then
         assertTrue(result.isPresent());
         assertEquals(enforcementOrderEntity, result.get());
-        verify(pcsCaseRepository).findByCaseReference(caseReference);
         verify(enforcementOrderRepository).findByClaimId(any(UUID.class));
     }
 
@@ -274,36 +183,29 @@ class EnforcementOrderMediatorTest {
     void shouldReturnEmptyWhenEnforcementOrderNotFound() {
         // Given
         PcsCaseEntity pcsCaseEntity = createPcsCaseEntity();
-
-        when(pcsCaseRepository.findByCaseReference(CASE_REFERENCE)).thenReturn(Optional.of(pcsCaseEntity));
         when(enforcementOrderRepository.findByClaimId(any(UUID.class))).thenReturn(List.of());
 
         // When
-        Optional<EnforcementOrderEntity> result = underTest.getEnforcementOrder(CASE_REFERENCE);
+        Optional<EnforcementOrderEntity> result = underTest.getEnforcementOrder(pcsCaseEntity);
 
         // Then
         assertTrue(result.isEmpty());
-        verify(pcsCaseRepository).findByCaseReference(CASE_REFERENCE);
         verify(enforcementOrderRepository).findByClaimId(any(UUID.class));
     }
 
     @Test
     void shouldNotSetMarkupWhenEnforcementOrderNotPresent() {
         // Given
-        long caseReference = 1234567890L;
         PcsCaseEntity pcsCaseEntity = createPcsCaseEntity();
 
-        when(pcsCaseRepository.findByCaseReference(caseReference))
-            .thenReturn(Optional.of(pcsCaseEntity));
         when(enforcementOrderRepository.findByClaimId(any(UUID.class))).thenReturn(List.of());
 
         // When
-        underTest.handleEnforcementRequirements(caseReference, pcsCase);
+        underTest.setCaseFields(pcsCase, pcsCaseEntity);
 
         // Then
         assertNull(pcsCase.getShowConfirmEvictionJourney());
         assertNull(pcsCase.getConfirmEvictionSummaryMarkup());
-        verify(pcsCaseRepository).findByCaseReference(caseReference);
         verify(enforcementOrderRepository).findByClaimId(any(UUID.class));
     }
